Fix typo in unexported function name
It looks like this was supposed to be "delim" write all -- "delitm"
doesn't mean anything.

Also skip casting the bytes slice to a string before formatting as %s,
as the cast is redundant; and prefer %w for wrapping an error, for
potential later use of errors.Is or errors.As.